requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.6.2
charset-normalizer==3.3.2
distro==1.9.0
filelock==3.14.0
fsspec==2024.5.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.2
idna==3.7
Jinja2==3.1.4
MarkupSafe==2.1.5
mlx==0.14.1
mlx-lm==0.14.3
mpmath==1.3.0
networkx==3.3
numpy==2.0.0rc2
openai==1.30.5
packaging==24.0
pillow==10.3.0
protobuf==5.27.0
pydantic==2.7.2
pydantic_core==2.18.3
python-dotenv==1.0.1
python-telegram-bot==21.2
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
safetensors==0.4.3
sniffio==1.3.1
sympy==1.12.1
tokenizers==0.19.1
torch==2.4.0.dev20240529
torchaudio==2.2.0.dev20240529
torchvision==0.19.0.dev20240529
tqdm==4.66.4
transformers==4.41.2
typing_extensions==4.12.1
urllib3==2.2.1
watchdog==4.0.1
